pyproject.toml
[project]
name = "cloud-engineer"
version = "0.1.0"
description = "Bedrock Agents / Strands based Agentic tooling test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.39.1",
    "diagrams>=0.24.4",
    "mcp>=1.10.1",
    "openpyxl>=3.1.5",
    "strands-agents>=0.1.9",
    "strands-agents-tools>=0.1.7",
    "streamlit>=1.46.1",
    "streamlit-mermaid>=0.3.0",
]
